backend/requirements.txt
fastapi==0.63.0
uvloop==0.15.2
httptools==0.1.1
SQLAlchemy==1.3.23
psycopg2-binary==2.8.6
uvicorn==0.13.4
gunicorn==20.0.4
aiofiles==0.6.0
requests==2.25.1
networkx==2.5
environs==9.3.1
